qc_engine/requirements.txt

# Minimal, no external subtitle libs required
azure-functions
lxml
